Enum recital::resolve::Operation

pub enum Operation {
    Exactly(Version),
    ExactlyNot(Version),
    GreaterThan(Version),
    GreaterThanOrEqualTo(Version),
    LessThan(Version),
    LessThanOrEqualTo(Version),
}

A version number constraint as an equality and inequality check.

The following example will create a very simple constraint based on an equality/inequality check that will allow any version number equal to or above 1.3.9.

use recital::Constraint;
use recital::Operation::*;
let operation = GreaterThanOrEqualTo(version!(1, 3, 9));

if operation.allows(&version!(1, 5, 6)) {
    // ... allowed ...
}

if !operation.allows(&version!(1, 2, 1)) {
    // ... not allowed ...
}

This enum is not intended to be used by itself. You would be better off doing direct comparisons (e.g. a > b). You are expected to use multiple instances of this enum as a set of constraints. Please see the Constraints enum documentation.

Variants

Exactly

Match the exact version number. (=)

ExactlyNot

Exclude the exact version number. (!=)

GreaterThan

Exclusively match any greater version number. (>)

GreaterThanOrEqualTo

Inclusively match any greater version number. (>=)

LessThan

Exclusively match any lesser version number. (<)

LessThanOrEqualTo

Inclusively match any lesser version number. (<=)