Module recital::version

Creates, parses, and manages semantic version numbers.

The version module provides you with a set of structs, enums, and macros to create a representation of a semantic version number. You have the option of using the structs and enums directly

use recital::version::{Identifier, Version};

let version = Version {
    major: 1,
    minor: 2,
    patch: 3,
    pre: vec![Identifier::Alpha("abc".to_string()),
              Identifier::Number(456)],
    build: vec![Identifier::Alpha("def".to_string()),
                Identifier::Number(789)],
};

or use the macros

let version = version!(1, 2, 3,
                       vec![id!("abc"), id!(456)],
                       vec![id!("def"), id!(789)]);

Structs

Version
Represents a semantic version number.

Enums

Identifier
Represents an alphanumeric or numberic identifier.