Enum recital::version::Identifier

pub enum Identifier {
    Number(u64),
    Alpha(String),
}

Represents an alphanumeric or numberic identifier.

An identifier in a semantic version number is a simple sequence of letters (a-z, A-Z), numbers (0-9), and hyphens that can appear in any order. Each identifier in a version number is separated by a dot (.). Knowing the difference between numeric and alphanumeric identifiers is important when version numbers are being compared.

Variants

Number

Reprents a numeric identifier.

A numeric identifier, as the name implies, may only have numbers. During comparison with other identifiers, numberc identifiers are compared normally to each other. However, when a numeric identifier is compared to a numeric identifier, it will always be the lesser of the two.

Alpha

Represents an alphanumeric identifier.

An alphanumeric identifier, as the name implies, may have letters and/or numbers. The letters are limited to a-z, A-Z, numbers to 0-9, and may also include one or more hyphens. During comparison with other identifiers, alphanumeric identifiers are compared normally to each other. However, when an alphanumeric identifier is compared to a numeric identifier, it will always be the greater of the two.