Struct recital::version::Version

pub struct Version {
    pub major: u64,
    pub minor: u64,
    pub patch: u64,
    pub pre: Vec<Identifier>,
    pub build: Vec<Identifier>,
}

Represents a semantic version number.

A semantic version number is a version number that conforms to a set of rules and requirements that have been defined in the Semantic Versioning specification. You may find the specification online at semver.org.

Fields

major

The major version number.

A major version number is used to indicate that a paritcular version is backward compatible with all other version numbers from the same major version number. For example, 1.999.999 is backwards compatible with 1.0.0, however 2.0.0 is not backwards compatible with 1.999.999.

minor

The minor version number.

A minor version number is used to indicate that one or more features have been added that are compatible with earlier versions of the same major version number.

patch

The patch version number.

A patch version number is used to indicate that one or more bugs have been fixed, but no new features have been added and compatibility with earlier versions of the same major version number is maintained.

pre

The pre-release identifiers.

The pre-release identifiers are used to indicate that this version number is not stable. While the identifiers can be as many as you want with whatever information you want, you will most commonly find people use identifiers such as -alpha, -beta.3, -rc.1, and so on. There may also be no identifiers, which indicates that this may be a stable version number.

build

The build identifiers.

The build identifiers do not mean anything in a semantic version number. While they are structured similarly (+debug, +xyz.123, etc.), they have no impact when comparing version numbers. They are mostly there to help developers identify other bits of information.

Methods

impl Version

fn clear_build(&mut self)

Removes all of the build identifiers.

fn clear_identifiers(&mut self)

Removes all of the pre-release and build identifiers.

fn clear_pre(&mut self)

Removes all of the pre-release identifiers.

fn increment_major(&mut self)

Increments the major version number.

When a major version number is incremented, the minor and patch version numbers are set to zero. In addition, the pre-release and build identifiers are all removed.

fn increment_minor(&mut self)

Increments the minor version number.

When a minor version number is incremented, the patch version number is set to zero. In addition, the pre-release and build identifiers are all removed.

fn increment_patch(&mut self)

Increments the patch version number.

When a patch version number is incremented, the pre-release and build identifiers are all removed.

fn is_stable(&self) -> bool

Checks if the version number is stable.

A version number is considered stable when it has a major version number that is greater than zero (0) and has no pre-release identifiers.

fn new() -> Version

Creates a new version number with default values.